pyproject.toml
[project]
name = "interview-preparation-agent"
version = "0.1.0"
description = "Interview Preparation Agent with A2A Protocol and LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk==0.3.2",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "langchain-openai >=0.1.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "langchain-community>=0.3.0",
    "duckduckgo-search>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"